Updating utilities for docstrings
Adding docstrings to mmctools. The path that is added needs to be a
relative path, but I'm not sure if it will be able to connect the
paths through GitHub... for example, it should probably be something
like '../mmctools/' since, from the docs/ dir perspective on GitHub,
mmctools is up one directory.